pyproject.toml
[project]
name = "banditdl"
version = "0.1.0"
description = "Bandit-ready decentralized learning experiments under Byzantine attacks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bayesianbandits>=1.3.0",
    "contextualbandits>=0.3.30",
    "hydra-core>=1.3.2",
    "mabwiser>=2.7.4",
    "matplotlib>=3.10.9",
    "networkx>=3.6.1",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pymab>=0.1.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "smpybandits>=0.9.7",
    "striatum>=0.2.5",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["banditdl"]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "mypy>=1.17.1",
    "pytest>=8.4.2",
    "ruff>=0.12.12",
]
